chore: move functions from integration test
commit-id:771a694f
