+++ crates/starknet_integration_tests/src/end_to_end_integration.rs
@@ -1,14 +1,8 @@
 use std::net::SocketAddr;
 
-use infra_utils::run_until::run_until;
-use infra_utils::tracing::{CustomLogger, TraceLevel};
 use mempool_test_utils::starknet_api_test_utils::{AccountId, MultiAccountTransactionGenerator};
-use papyrus_execution::execution_utils::get_nonce_at;
-use papyrus_storage::state::StateStorageReader;
-use papyrus_storage::StorageReader;
 use starknet_api::block::BlockNumber;
-use starknet_api::core::{ContractAddress, Nonce};
-use starknet_api::state::StateNumber;
+use starknet_api::core::Nonce;
 use starknet_sequencer_infra::test_utils::AvailablePorts;
 use starknet_sequencer_node::config::component_config::ComponentConfig;
 use starknet_sequencer_node::config::component_execution_config::{
@@ -20,6 +14,7 @@ use starknet_types_core::felt::Felt;
 use tracing::info;
 
 use crate::integration_test_setup::IntegrationTestSetup;
+use crate::sequencer_manager::{await_block, get_account_nonce};
 use crate::test_identifiers::TestIdentifier;
 use crate::utils::send_account_txs;
 
@@ -27,46 +22,6 @@ use crate::utils::send_account_txs;
 const N_CONSOLIDATED_SEQUENCERS: usize = 3;
 /// The number of distributed remote sequencers that participate in the test.
 const N_DISTRIBUTED_SEQUENCERS: usize = 2;
-
-/// Reads the latest block number from the storage.
-fn get_latest_block_number(storage_reader: &StorageReader) -> BlockNumber {
-    let txn = storage_reader.begin_ro_txn().unwrap();
-    txn.get_state_marker()
-        .expect("There should always be a state marker")
-        .prev()
-        .expect("There should be a previous block in the storage, set by the test setup")
-}
-
-/// Reads an account nonce after a block number from storage.
-fn get_account_nonce(storage_reader: &StorageReader, contract_address: ContractAddress) -> Nonce {
-    let block_number = get_latest_block_number(storage_reader);
-    let txn = storage_reader.begin_ro_txn().unwrap();
-    let state_number = StateNumber::unchecked_right_after_block(block_number);
-    get_nonce_at(&txn, state_number, None, contract_address)
-        .expect("Should always be Ok(Some(Nonce))")
-        .expect("Should always be Some(Nonce)")
-}
-
-/// Sample a storage until sufficiently many blocks have been stored. Returns an error if after
-/// the given number of attempts the target block number has not been reached.
-async fn await_block(
-    interval: u64,
-    target_block_number: BlockNumber,
-    max_attempts: usize,
-    storage_reader: &StorageReader,
-) -> Result<BlockNumber, ()> {
-    let condition = |&latest_block_number: &BlockNumber| latest_block_number >= target_block_number;
-    let get_latest_block_number_closure = || async move { get_latest_block_number(storage_reader) };
-
-    let logger = CustomLogger::new(
-        TraceLevel::Info,
-        Some("Waiting for storage to include block".to_string()),
-    );
-
-    run_until(interval, max_attempts, get_latest_block_number_closure, condition, Some(logger))
-        .await
-        .ok_or(())
-}
 
 pub async fn end_to_end_integration(tx_generator: MultiAccountTransactionGenerator) {
     const EXPECTED_BLOCK_NUMBER: BlockNumber = BlockNumber(15);
