+++ crates/starknet_integration_tests/src/lib.rs
@@ -2,6 +2,7 @@ pub mod config_utils;
 pub mod end_to_end_integration;
 pub mod flow_test_setup;
 pub mod integration_test_setup;
+pub mod sequencer_manager;
 pub mod state_reader;
 pub mod test_identifiers;
 pub mod utils;
