+++ crates/starknet_integration_tests/src/sequencer_manager.rs
@@ -1,0 +1,51 @@
+use infra_utils::run_until::run_until;
+use infra_utils::tracing::{CustomLogger, TraceLevel};
+use papyrus_execution::execution_utils::get_nonce_at;
+use papyrus_storage::state::StateStorageReader;
+use papyrus_storage::StorageReader;
+use starknet_api::block::BlockNumber;
+use starknet_api::core::{ContractAddress, Nonce};
+use starknet_api::state::StateNumber;
+
+/// Reads the latest block number from the storage.
+pub fn get_latest_block_number(storage_reader: &StorageReader) -> BlockNumber {
+    let txn = storage_reader.begin_ro_txn().unwrap();
+    txn.get_state_marker()
+        .expect("There should always be a state marker")
+        .prev()
+        .expect("There should be a previous block in the storage, set by the test setup")
+}
+
+/// Reads an account nonce after a block number from storage.
+pub fn get_account_nonce(
+    storage_reader: &StorageReader,
+    contract_address: ContractAddress,
+) -> Nonce {
+    let block_number = get_latest_block_number(storage_reader);
+    let txn = storage_reader.begin_ro_txn().unwrap();
+    let state_number = StateNumber::unchecked_right_after_block(block_number);
+    get_nonce_at(&txn, state_number, None, contract_address)
+        .expect("Should always be Ok(Some(Nonce))")
+        .expect("Should always be Some(Nonce)")
+}
+
+/// Sample a storage until sufficiently many blocks have been stored. Returns an error if after
+/// the given number of attempts the target block number has not been reached.
+pub async fn await_block(
+    interval: u64,
+    target_block_number: BlockNumber,
+    max_attempts: usize,
+    storage_reader: &StorageReader,
+) -> Result<BlockNumber, ()> {
+    let condition = |&latest_block_number: &BlockNumber| latest_block_number >= target_block_number;
+    let get_latest_block_number_closure = || async move { get_latest_block_number(storage_reader) };
+
+    let logger = CustomLogger::new(
+        TraceLevel::Info,
+        Some("Waiting for storage to include block".to_string()),
+    );
+
+    run_until(interval, max_attempts, get_latest_block_number_closure, condition, Some(logger))
+        .await
+        .ok_or(())
+}
